Implement proper navigation to the ReviewersScreen.kt.

+++ app/src/main/java/com/appunite/loudius/domain/model/Reviewer.kt
@@ -4,6 +4,6 @@ data class Reviewer(
     val id: Int,
     val login: String,
     val isReviewDone: Boolean,
-    val hoursFromPRStart: Int,
-    val hoursFromReviewDone: Int?,
+    val hoursFromPRStart: Long,
+    val hoursFromReviewDone: Long?,
 )
